+++ lib/gangplank.dart
@@ -8,8 +8,8 @@ import 'package:gangplank/src/storage.dart';
 
 export './src/lcu_watcher.dart' show LCUWatcher, LCUWatcherConfig, LCUCredentials;
 export './src/lcu_socket.dart' show LCUSocket, LCUSocketConfig, EventResponse, ManualEventResponse;
-export './src/lcu_http_client.dart' show LCUHttpClient, LCUHttpClientConfig, LCUHttpClientException, LCUGetRouteToCache, LCUGetRouteToCacheMatchType;
-export './src/lcu_live_game_watcher.dart' show LCULiveGameWatcher, LCULiveGameWatcherConfig, LCULiveGameWatcherSummary;
+export './src/lcu_http_client.dart' show LCUHttpClient, LCUHttpClientConfig, LCUHttpClientException, LCUGetRouteToCache;
+export './src/lcu_live_game_watcher.dart' show LCULiveGameWatcher, LCULiveGameWatcherConfig, LCULiveGameWatcherSummary, GamePresenceCheckStrategy;
 
 class GangplankHttpOverrides extends HttpOverrides {
   @override